clean up, ... thx gui-dos for your tipp
app badge for mmol values

+++ LibreDirectLibrary/Content/GlucoseUnit.swift
@@ -2,7 +2,7 @@
 //  GlucoseUnit.swift
 //  LibreDirect
 //
-//  Created by Reimar Metzen on 02.08.21. 
+//  Created by Reimar Metzen on 02.08.21.
 //
 
 import Foundation
@@ -11,7 +11,7 @@ public enum GlucoseUnit: String, Codable {
     case mgdL = "mg/dL"
     case mmolL = "mmol/L"
 
-    static let exchangeRate: Decimal = 0.0555
+    static let exchangeRate: Double = 0.0555
     
     public var description: String {
         switch self {
